+++ apps/connect/src/pages/content.tsx
@@ -5,12 +5,15 @@ import { useNodeNavigation } from '#hooks/useNodeNavigation';
 import { useUiNodes } from '#hooks/useUiNodes';
 import { FILE } from '@powerhousedao/design-system';
 import { useEffect } from 'react';
+import { DocumentEditorContainer } from 'src/components/document-editor-container';
+import { DriveView } from 'src/components/drive-view';
+import { Footer } from 'src/components/footer';
+import { useUiNodes } from 'src/hooks/useUiNodes';
 
 export default function Content() {
     const uiNodes = useUiNodes();
     const { fileNodeDocument, selectedDriveNode, selectedNode, addFile } =
         uiNodes;
-    useNodeNavigation();
 
     useEffect(() => {
         return window.electronAPI?.handleFileOpen(async file => {