+++ clis/ph-cli/src/commands/connect.ts
@@ -1,11 +1,14 @@
+import { getConfig } from "@powerhousedao/config/powerhouse";
 import {
   ConnectStudioOptions,
   startConnectStudio,
 } from "@powerhousedao/builder-tools/connect-studio";
 import { getConfig } from "@powerhousedao/config/powerhouse";
 import { Command } from "commander";
+import { version } from "../../package.json";
+export type ConnectOptions = ConnectStudioOptions;
 
-export async function startConnect(connectOptions: ConnectStudioOptions) {
+export async function startConnect(connectOptions: ConnectOptions) {
   const { documentModelsDir, editorsDir } = getConfig();
   const options = { documentModelsDir, editorsDir, ...connectOptions };
   return await startConnectStudio(options);
@@ -34,10 +37,10 @@ export function connectCommand(program: Command) {
     .action(async (...args: [ConnectStudioOptions]) => {
       const connectOptions = args.at(0) || {};
       const { documentModelsDir, editorsDir, packages, studio } = getConfig();
-
       await startConnectStudio({
         port: studio?.port?.toString() || undefined,
         packages,
+        phCliVersion: typeof version === "string" ? version : undefined,
         localDocuments: documentModelsDir || undefined,
         localEditors: editorsDir || undefined,
         open: studio?.openBrowser,