+++ packages/design-system/src/connect/components/home-screen/index.ts
@@ -1,0 +1,3 @@
+export * from "./home-screen";
+export * from "./home-screen-add-drive-item";
+export * from "./home-screen-item";