+++ packages/design-system/src/connect/components/modal/settings-modal-v2/about.stories.tsx
@@ -15,3 +15,10 @@ export const Default: Story = {
     packageJson: mockPackageJson,
   },
 };
+
+export const WithPhCliVersion: Story = {
+  args: {
+    packageJson: mockPackageJson,
+    phCliVersion: "1.0.0",
+  },
+};