+++ packages/design-system/src/connect/components/modal/settings-modal/dependency-versions/dependency-versions.stories.tsx
@@ -23,3 +23,17 @@ export const Default: Story = {
     packageJson: mockPackageJson,
   },
 };
+
+export const WithPhCliVersion: Story = {
+  decorators: [
+    (Story) => (
+      <div className="w-[320px]">
+        <Story />
+      </div>
+    ),
+  ],
+  args: {
+    packageJson: mockPackageJson,
+    phCliVersion: "1.0.0",
+  },
+};