+++ packages/design-system/src/powerhouse/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./animation";
 export * from "./usePagination";
+export * from "./useWindowSize";